refactor: dto 클래스들을 별도 디렉토리로 분리
