+++ src/main/java/com/github/thundermarket/thundermarket/dto/ProductDetailResponse.java
@@ -1,4 +1,6 @@
-package com.github.thundermarket.thundermarket.domain;
+package com.github.thundermarket.thundermarket.dto;
+
+import com.github.thundermarket.thundermarket.domain.ProductDetail;
 
 import java.util.Objects;
