+++ src/main/java/com/github/thundermarket/thundermarket/repository/NCloudStorage.java
@@ -1,7 +1,7 @@
 package com.github.thundermarket.thundermarket.repository;
 
 import com.github.thundermarket.thundermarket.Util.VideoUtils;
-import com.github.thundermarket.thundermarket.domain.FileUploadResult;
+import com.github.thundermarket.thundermarket.dto.FileUploadResult;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 import com.github.thundermarket.thundermarket.constant.*;
