+++ src/test/java/com/github/thundermarket/thundermarket/testDouble/ProductFakeRepository.java
@@ -1,7 +1,7 @@
 package com.github.thundermarket.thundermarket.testDouble;
 
 import com.github.thundermarket.thundermarket.domain.Product;
-import com.github.thundermarket.thundermarket.domain.ProductFilterRequest;
+import com.github.thundermarket.thundermarket.dto.ProductFilterRequest;
 import com.github.thundermarket.thundermarket.repository.ProductRepository;
 import org.springframework.data.domain.Pageable;
